+++ iopipe/iopipe.py
@@ -1,12 +1,8 @@
-import datetime
 import json
-import platform
-import socket
 import time
 import os
-import traceback
 
-import constants
+from report import Report
 from collector import get_collector_url
 
 try:
@@ -40,147 +36,6 @@ class IOpipe(object):
         self._url = url
         self._debug = debug
         self.client_id = client_id
-        self.report = {
-            'client_id': self.client_id,
-            'environment': {
-                'host': {},
-                'os': {
-                    'cpus': [],
-                    'linux': {
-                        'mem': {},
-                        'pid': {
-                            'self': {}
-                        }
-                    }
-                },
-                'python': {}
-            },
-            "custom_metrics": []
-        }
-
-        self.report['environment']['os']['linux']['pid'].update({
-            'self': {
-                'stat_start': get_pid_stat('self')
-            }
-        })
-
-    def __del__(self):
-        """
-        Send the report if it hasn't already been sent
-        """
-        self.send()
-
-    def _add_aws_lambda_data(self, context):
-        """
-        Add AWS Lambda specific data to the report
-        """
-        aws_key = 'aws'
-        self.report[aws_key] = {}
-
-        for k, v in {
-            # camel case names in the report to align with AWS standards
-            'functionName': 'function_name',
-            'functionVersion': 'function_version',
-            'memoryLimitInMB': 'memory_limit_in_mb',
-            'invokedFunctionArn': 'invoked_function_arn',
-            'awsRequestId': 'aws_request_id',
-            'logGroupName': 'log_group_name',
-            'logStreamName': 'log_stream_name',
-        }.items():
-            if v in dir(context):
-                self.report[aws_key][k] = getattr(context, v)
-
-        if context and 'get_remaining_time_in_millis' in dir(context):
-            try:
-                self.report['aws']['getRemainingTimeInMillis'] = \
-                    context.get_remaining_time_in_millis()
-            except Exception:
-                pass  # @TODO handle this more gracefully
-
-    def _add_pid_data(self, pid):
-        self.report['environment']['os']['linux']['pid'][pid] = \
-            self.report['environment']['os']['linux']['pid'][pid] or {}
-
-        self.report['environment']['os']['linux']['pid'][pid]['stat'] = \
-            get_pid_stat(pid)
-
-        with open("/proc/%s/status" % (pid,)) as status_file:
-            status = {}
-            for row in status_file:
-                line = row.split(":")
-                status_value = line[1].rstrip("\t\n kB").lstrip()
-                try:
-                    status[line[0]] = int(status_value)
-                except ValueError:
-                    status[line[0]] = status_value
-            self.report['environment']['os']['linux']['pid'][pid]['status'] \
-                = status
-
-    def _add_os_host_data(self):
-        """
-        Add os field to payload
-        """
-        uptime = None
-
-        with open("/proc/stat") as stat_file:
-            for line in stat_file:
-                cpu_stat = line.split(" ")
-                if cpu_stat[0][:3] != "cpu":
-                    break
-                # First cpu line is aggregation of following lines, skip it
-                if len(cpu_stat[0]) == 3:
-                    continue
-                self.report['environment']['os']['cpus'].append({
-                    'name': cpu_stat[0],
-                    'times': {
-                        'user': cpu_stat[1],
-                        'nice': cpu_stat[2],
-                        'sys': cpu_stat[3],
-                        'idle': cpu_stat[4],
-                        'irq': cpu_stat[6]
-                    }
-                })
-
-        with open("/proc/uptime") as uptime_file:
-            utf = uptime_file.readline().split(" ")
-            uptime = int(float(utf[0]))
-
-        with open("/proc/sys/kernel/random/boot_id") as bootid_file:
-            self.report['environment']['host'].update({
-                'container_id': bootid_file.readline()
-            })
-
-        with open("/proc/meminfo") as meminfo:
-            linux_mem = {}
-            for row in meminfo:
-                line = row.split(":")
-                # Example content:
-                # MemTotal:                3801016 kB
-                # MemFree:                 1840972 kB
-                # MemAvailable:        3287752 kB
-                # HugePages_Total:             0
-                linux_mem[line[0]] = \
-                    int(line[1].lstrip().rstrip(" kB\n"))
-            self.report['environment']['os']['linux']['mem'].update(linux_mem)
-
-        self.report['environment']['os'].update({
-            'hostname': socket.gethostname(),
-            'uptime': uptime,
-            'freemem':
-                self.report['environment']['os']['linux']['mem']['MemFree'],
-            'totalmem':
-                self.report['environment']['os']['linux']['mem']['MemTotal'],
-            'usedmem':
-                self.report['environment']['os']['linux']['mem']['MemTotal'] -
-                self.report['environment']['os']['linux']['mem']['MemFree']
-        })
-
-    def _add_python_local_data(self):
-        """
-        Add the python sys attributes relevant to AWS Lambda execution
-        """
-        self.report['environment']['python']['version'] = \
-            platform.python_version()
 
     def log(self, key, value):
         """
@@ -197,63 +52,18 @@ class IOpipe(object):
             event['n'] = value
         else:
             event['s'] = str(value)
-        self.report['custom_metrics'].append(event)
+        self.report.custom_metrics.append(event)
 
-    def err(self, err):
-        """
-        Add the details of an error to the report
-        """
-        err_details = {
-            'name': type(err).__name__,
-            'message': '{}'.format(err),
-            'stack': traceback.format_exc(),
-            'time_reported': datetime.datetime.now()
-            .strftime(constants.TIMESTAMP_FORMAT)
-        }
-        if 'errors' not in self.report:
-            self.report['errors'] = err_details
-
-        self._add_python_local_data()
-
-    def send(self, context=None, time_start=None):
+    def send(self, report, time_start=None):
         """
         Send the current report to IOpipe
         """
-        global COLDSTART
         json_report = None
 
-        self._add_pid_data('self')
-
-        # Duration of execution.
-        duration = time.time() - (time_start or time.time())
-        self.report.update({
-            'duration': int(duration * 1000000000),
-            'time_sec': int(duration),
-            'time_nanosec': int((duration - int(duration)) * 1000000000),
-            'coldstart': COLDSTART
-        })
-
-        if COLDSTART:
-            COLDSTART = False
-
-        self.report['environment'].update(
-            {
-                'agent': {
-                  'runtime': "python",
-                  'version': constants.VERSION,
-                  'load_time': MODULE_LOAD_TIME
-                }
-            })
-
-        if context:
-            self._add_aws_lambda_data(context)
-        self._add_python_local_data()
-        self._add_os_host_data()
-
         try:
-            json_report = json.dumps(self.report)
+            json_report = json.dumps(report, default=lambda o: o.__dict__)
         except Exception as err:
-            print("Could not convert the report to JSON."
+            print("Could not convert the report to JSON. "
                   "Threw exception: {}".format(err))
             print('Report: {}'.format(self.report))
             return
@@ -270,21 +80,26 @@ class IOpipe(object):
         finally:
             if self._debug:
                 print(json_report)
-            # Clear custom metrics & error data between invocations!
-            self.report['custom_metrics'] = []
-            if 'errors' in self.report:
-                del self.report['errors']
 
     def decorator(self, fun):
         def wrapped(event, context):
+            global COLDSTART
             err = None
             start_time = time.time()
+            invocation_report = Report(self.client_id,
+                                       get_pid_stat('self'),
+                                       MODULE_LOAD_TIME)
+            self.report = invocation_report
+            self.start_time = start_time
             try:
                 result = fun(event, context)
+                invocation_report.update_data(context, COLDSTART, start_time)
+                COLDSTART = False
             except Exception as err:
-                self.err(err)
+                invocation_report.retain_err(err)
+                self.send(self.report, self.start_time)
                 raise err
             finally:
-                self.send(context, start_time)
+                self.send(invocation_report, start_time)
             return result
         return wrapped
