+++ iopipe/tests/MockContext.py
@@ -2,7 +2,7 @@ class MockContext(object):
     function_version = '$LATEST'
     aws_request_id = 0
     log_group_name = 'mock-group'
-    logs_stream_name = 'mock-stream'
+    log_stream_name = 'mock-stream'
     memory_limit_in_mb = 500
 
     def __init__(self, name='handler', version='$LATEST'):
